+++ app/src/main/java/com/nohjason/minari/network/ApiService.kt
@@ -13,13 +13,13 @@ import com.nohjason.minari.screens.login.response.LoginRequest
 import com.nohjason.minari.screens.login.response.LoginResponse
 import com.nohjason.minari.screens.login.response.RegisterRequest
 import com.nohjason.minari.screens.login.response.RegisterResponse
-import com.nohjason.minari.screens.profile.DirecGpResponse
-import com.nohjason.minari.screens.profile.DirecGpsResponse
-import com.nohjason.minari.screens.profile.DirecGpseResponse
-import com.nohjason.minari.screens.profile.DirecTermResponse
+import com.nohjason.minari.screens.profile.directory.DirecGpResponse
+import com.nohjason.minari.screens.profile.directory.DirecGpsResponse
+import com.nohjason.minari.screens.profile.directory.DirecGpseResponse
+import com.nohjason.minari.screens.profile.directory.DirecTermResponse
 import com.nohjason.minari.screens.rout.response.GetAllNews
 import com.nohjason.minari.screens.rout.response.LikesResponse
-import com.nohjason.minari.screens.profile.ProfileResponse
+import com.nohjason.minari.screens.profile.profile_data.ProfileResponse
 import com.nohjason.minari.screens.quiz.data.QuestionResponse
 import com.nohjason.myapplication.network.response.Term
 import com.nohjason.myapplication.network.response.TermResponse
@@ -137,27 +137,27 @@ interface ApiService {
 
     @GET("/questions/level/{level}")
     suspend fun getQuestion(
-        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjYxMzY2MzgsImV4cCI6MTcyNjIyMzAzOH0.kUqZInuvOuSvxmL_kRXSBoJ725F5URKufOy0L7QRWBM",
+        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcyNzQ2NDksImV4cCI6MTcyNzM2MTA0OX0.7j-9kpq7s7tJJcpM41Q_Nw3BJx2JIYwQRkCnoyqjntY",
         @Path("level") level: Int
     ): QuestionResponse
 
     @GET("/likes/term")
     suspend fun getDiercTerm(
-        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcwNzQyODAsImV4cCI6MTcyNzE2MDY4MH0._PEUQnz1VtDZdX_0KpiPVVso6uUeFPw6YuEMUxU8Stk",
+        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcyNzQ2NDksImV4cCI6MTcyNzM2MTA0OX0.7j-9kpq7s7tJJcpM41Q_Nw3BJx2JIYwQRkCnoyqjntY",
     ): DirecTermResponse
 
     @GET("/likes/gpse")
     suspend fun getGpse(
-        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcwNzQyODAsImV4cCI6MTcyNzE2MDY4MH0._PEUQnz1VtDZdX_0KpiPVVso6uUeFPw6YuEMUxU8Stk",
+        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcyNzQ2NDksImV4cCI6MTcyNzM2MTA0OX0.7j-9kpq7s7tJJcpM41Q_Nw3BJx2JIYwQRkCnoyqjntY",
     ): DirecGpseResponse
 
     @GET("/likes/gps")
     suspend fun getGps(
-        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcwNzQyODAsImV4cCI6MTcyNzE2MDY4MH0._PEUQnz1VtDZdX_0KpiPVVso6uUeFPw6YuEMUxU8Stk",
+        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcyNzQ2NDksImV4cCI6MTcyNzM2MTA0OX0.7j-9kpq7s7tJJcpM41Q_Nw3BJx2JIYwQRkCnoyqjntY",
     ): DirecGpsResponse
 
     @GET("/likes/gp")
     suspend fun getGp(
-        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcwNzQyODAsImV4cCI6MTcyNzE2MDY4MH0._PEUQnz1VtDZdX_0KpiPVVso6uUeFPw6YuEMUxU8Stk",
+        @Header("Authorization") Authorization: String = "eyJhbGciOiJIUzI1NiJ9.eyJhdXRob3JpdHkiOiJST0xFX1VTRVIiLCJzdWIiOiJyaGRpZGRsNjY5MUBnbWFpbC5jb20iLCJpYXQiOjE3MjcyNzQ2NDksImV4cCI6MTcyNzM2MTA0OX0.7j-9kpq7s7tJJcpM41Q_Nw3BJx2JIYwQRkCnoyqjntY",
     ): DirecGpResponse
 }
