+++ app/src/main/java/com/nohjason/minari/network/MainViewModel.kt
@@ -3,22 +3,11 @@ package com.nohjason.myapplication.network
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nohjason.minari.network.response.BookResponse
-import com.nohjason.minari.screens.profile.ProfileResponse
 import com.nohjason.myapplication.network.RetrofitInstance.api
 import com.nohjason.myapplication.network.response.Term
-import com.nohjason.myapplication.network.response.TermResponse
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import retrofit2.HttpException
-import java.io.IOException
-import java.net.SocketTimeoutException
 
 class MainViewModel : ViewModel() {
