+++ app/src/main/java/com/nohjason/minari/screens/profile/ProfileMainScreen.kt
@@ -25,10 +25,8 @@ import androidx.navigation.NavHostController
 import com.nohjason.minari.R
 import com.nohjason.minari.preferences.getFromPreferences
 import com.nohjason.minari.preferences.getPreferences
-import com.nohjason.minari.navigation.bottombar.BottomScreen
-import com.nohjason.minari.screens.profile.LikeList
-import com.nohjason.minari.screens.profile.ProfileResponse
-import com.nohjason.minari.screens.profile.ProfileViewModel
+import com.nohjason.minari.screens.profile.profile_data.ProfileResponse
+import com.nohjason.minari.screens.profile.profile_data.ProfileViewModel
 import com.nohjason.minari.screens.profile.element.ProfileButton
 import com.nohjason.minari.screens.profile.element.ProfileInfor
 import com.nohjason.minari.screens.profile.element.RewardBar
