+++ app/src/main/java/com/nohjason/minari/screens/profile/alias/AliasMainCard.kt
@@ -22,14 +22,14 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import com.nohjason.minari.screens.profile.WebLink
+import com.nohjason.minari.screens.profile.profile_data.WebLinkData
 
 @Composable
 fun AliasMainCard(
     level: Int,
     exp: Int,
 ){
-    val webLink = WebLink()
+    val webLink = WebLinkData()
     val getWebLink = webLink.getTitleAndUrlForLevel(level)
     val progress = (exp / 100.toFloat()).coerceIn(0f, 1f)
     Box(
