+++ app/src/main/java/com/nohjason/minari/screens/profile/alias/AliasScreen.kt
@@ -11,7 +11,9 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.material3.Divider
@@ -26,15 +28,14 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.nohjason.minari.R
-import com.nohjason.minari.screens.profile.likes.Like
-import com.nohjason.minari.screens.quiz.quiz_end.TitleBar
+import com.nohjason.minari.screens.ui.titlebar.TitleBar
 
 @Composable
 fun AliasScreen(
     level: Int,
     exp: Int
 ){
-
+    val scrollState = rememberScrollState()
     val nameList = remember {
         List(30) { index -> "Title $index" }
     }
@@ -42,7 +43,8 @@ fun AliasScreen(
     Column (
         modifier = Modifier
             .fillMaxSize()
-            .background(Color(0xFFF5F6FA)),
+            .background(Color(0xFFF5F6FA))
+            .verticalScroll(scrollState),
         horizontalAlignment = Alignment.CenterHorizontally
     ){
         TitleBar(
