+++ app/src/main/java/com/nohjason/minari/screens/profile/profile_data/TestDummyData.kt
@@ -1,6 +1,13 @@
-package com.nohjason.minari.screens.profile
+package com.nohjason.minari.screens.profile.profile_data
 
 import com.nohjason.minari.screens.profile.directory.DirecGp
+import com.nohjason.minari.screens.profile.directory.DirecGpResponse
+import com.nohjason.minari.screens.profile.directory.DirecGps
+import com.nohjason.minari.screens.profile.directory.DirecGpsResponse
+import com.nohjason.minari.screens.profile.directory.DirecGpse
+import com.nohjason.minari.screens.profile.directory.DirecGpseResponse
+import com.nohjason.minari.screens.profile.directory.DirecTerm
+import com.nohjason.minari.screens.profile.directory.DirecTermResponse
 
 // ProfileData 더미 데이터
 object DummyProfileData {
@@ -24,19 +31,19 @@ object DummyTermStatusResponse {
         status = 0,
         message = "Success",
         data = listOf(
-            Term(
+            DirecTerm(
                 termId = 1,
                 termNm = "Inflation",
                 termExplain = "The rate at which the general level of prices for goods and services is rising.",
                 termDifficulty = TermDifficulty.LV_1
             ),
-            Term(
+            DirecTerm(
                 termId = 2,
                 termNm = "Supply and Demand",
                 termExplain = "The relationship between the quantity of a commodity that producers wish to sell at various prices and the quantity that consumers wish to buy.",
                 termDifficulty = TermDifficulty.LV_2
             ),
-            Term(
+            DirecTerm(
                 termId = 3,
                 termNm = "Opportunity Cost",
                 termExplain = "The loss of potential gain from other alternatives when one alternative is chosen.",
@@ -52,13 +59,13 @@ object DummyGpseStatusResponse {
         status = 0,
         message = "Success",
         data = listOf(
-            Gpse(
+            DirecGpse(
                 gpseId = 1,
                 gpseName = "Gpse Example 1",
                 gpseTime = 120,
                 gpseLike = true
             ),
-            Gpse(
+            DirecGpse(
                 gpseId = 2,
                 gpseName = "Gpse Example 2",
                 gpseTime = 45,
@@ -74,14 +81,14 @@ object DummyGpsStatusResponse {
         status = 0,
         message = "Success",
         data = listOf(
-            Gps(
+            DirecGps(
                 gpsId = 1,
                 gpsContent = "Gps Content Example 1",
                 gpsImg = "https://i.ibb.co/kJPp4Cv/image-34.png",
                 gpsLike = true,
                 gpsTpList = listOf("BEGINNER", "INTERMEDIATE")
             ),
-            Gps(
+            DirecGps(
                 gpsId = 2,
                 gpsContent = "Gps Content Example 2",
                 gpsImg = "https://i.ibb.co/kJPp4Cv/image-34.png",
@@ -98,13 +105,13 @@ object DummyGpStatusResponse {
         status = 0,
         message = "Success",
         data = listOf(
-            Gp(
+            DirecGp(
                 gpId = 1,
                 gpName = "Gp Example 1",
                 gpImg = "https://i.ibb.co/kJPp4Cv/image-34.png",
                 gpLike = true
             ),
-            Gp(
+            DirecGp(
                 gpId = 2,
                 gpName = "Gp Example 2",
                 gpImg = "https://i.ibb.co/kJPp4Cv/image-34.png",
